+++ src/refactoring/widgets/index.ts
@@ -1,0 +1,2 @@
+export { ProductList } from './ProductList';
+export { CartList } from './CartList';
